+++ models/experimental/stable_diffusion_xl_base/tests/test_sdxl_base_refiner_accuracy.py
@@ -3,32 +3,24 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import pytest
-import csv
 from models.experimental.stable_diffusion_xl_base.demo.demo_base_and_refiner import test_demo_base_and_refiner
-from models.experimental.stable_diffusion_xl_base.utils.clip_encoder import CLIPEncoder
 import os
-import urllib
 from loguru import logger
-import statistics
-from models.experimental.stable_diffusion_xl_base.utils.fid_score import calculate_fid_score
 from models.experimental.stable_diffusion_xl_base.tests.test_common import (
     SDXL_L1_SMALL_SIZE,
     SDXL_BASE_REFINER_TRACE_REGION_SIZE,
     SDXL_FABRIC_CONFIG,
 )
 import json
-from models.common.utility_functions import profiler
 from models.experimental.stable_diffusion_xl_base.conftest import get_device_name
-from models.experimental.stable_diffusion_xl_base.utils.clip_fid_ranges import (
-    accuracy_check_clip,
-    accuracy_check_fid,
-    get_appr_delta_metric,
-    get_model_targets,
+from models.experimental.stable_diffusion_xl_base.utils.accuracy_utils import (
+    sdxl_get_prompts,
+    calculate_accuracy_metrics,
+    create_report_json,
+    save_report_json,
 )
 
 test_demo_base_and_refiner.__test__ = False
-COCO_CAPTIONS_DOWNLOAD_PATH = "https://github.com/mlcommons/inference/raw/4b1d1156c23965172ae56eacdd8372f8897eb771/text_to_image/coco2014/captions/captions_source.tsv"
-OUT_ROOT, RESULTS_FILE_NAME = "test_reports", "sdxl_test_results.json"
 
 
 @pytest.mark.parametrize(
@@ -171,150 +163,26 @@ def test_accuracy_sdxl_base_and_refiner(
     if skip_check_and_save:
         logger.info("Skipping accuracy check and saving results as per environment variable.")
         return
-    clip = CLIPEncoder()
 
-    clip_scores = []
+    accuracy_metrics = calculate_accuracy_metrics(images, prompts, coco_statistics_path)
 
-    for idx, image in enumerate(images):
-        clip_scores.append(100 * clip.get_clip_score(prompts[idx], image).item())
-
-    average_clip_score = sum(clip_scores) / len(clip_scores)
-
-    deviation_clip_score = "N/A"
-    fid_score = "N/A"
-
-    if num_prompts >= 2:
-        deviation_clip_score = statistics.stdev(clip_scores)
-        fid_score = calculate_fid_score(images, coco_statistics_path)
-    else:
-        logger.info("FID score is not calculated for less than 2 prompts.")
-
-    avg_gen_end_to_end = profiler.get("end_to_end_generation")
-    model_name = "sdxl-base-refiner-tp" if use_cfg_parallel else "sdxl-base-refiner"
-
-    targets = get_model_targets(model_name)
-
-    data = {
-        "model": model_name,
-        "metadata": {
-            "model_name": model_name,
-            "device": get_device_name(),
-            "device_vae": vae_on_device,
-            "capture_trace": capture_trace,
-            "encoders_on_device": encoders_on_device,
-            "num_inference_steps": num_inference_steps,
-            "start_from": start_from,
-            "num_prompts": num_prompts,
-            "negative_prompt": negative_prompt,
-            "guidance_scale": guidance_scale,
-            "use_refiner": use_refiner,
-            "denoising_split": denoising_split,
-            "refiner_strength": refiner_strength,
-            "refiner_aesthetic_score": refiner_aesthetic_score,
-            "refiner_negative_aesthetic_score": refiner_negative_aesthetic_score,
-        },
-        "benchmarks_summary": [
-            {
-                "model": model_name,
-                "device": get_device_name(),
-                "avg_gen_time": avg_gen_end_to_end,
-                "target_checks": {
-                    "functional": {
-                        "avg_gen_time": targets["perf"]["functional"],
-                        "avg_gen_time_check": 2 if targets["perf"]["functional"] >= avg_gen_end_to_end else 3,
-                    },
-                    "complete": {
-                        "avg_gen_time": targets["perf"]["complete"],
-                        "avg_gen_time_check": 2 if targets["perf"]["complete"] >= avg_gen_end_to_end else 3,
-                    },
-                    "target": {
-                        "avg_gen_time": targets["perf"]["target"],
-                        "avg_gen_time_check": 2 if targets["perf"]["target"] >= avg_gen_end_to_end else 3,
-                    },
-                },
-                "average_denoising_time": profiler.get("denoising_loop"),
-                "average_vae_time": profiler.get("vae_decode"),
-                "min_gen_time": min(profiler.times["end_to_end_generation"]),
-                "max_gen_time": max(profiler.times["end_to_end_generation"]),
-                "average_encoding_time": profiler.get("encode_prompts"),
-            }
-        ],
-        "evals": [
-            {
-                "model": model_name,
-                "device": get_device_name(),
-                "average_clip": average_clip_score,
-                "deviation_clip": deviation_clip_score,
-                "clip_accuracy_check_approx": accuracy_check_clip(
-                    model_name, average_clip_score, num_prompts, mode="approx"
-                ),
-                "clip_accuracy_check_valid": accuracy_check_clip(
-                    model_name, average_clip_score, num_prompts, mode="valid"
-                ),
-                "delta_clip": get_appr_delta_metric(model_name, average_clip_score, num_prompts, score_type="clip"),
-                "fid_score": fid_score,
-                "fid_accuracy_check_approx": accuracy_check_fid(model_name, fid_score, num_prompts, mode="approx"),
-                "fid_accuracy_check_valid": accuracy_check_fid(model_name, fid_score, num_prompts, mode="valid"),
-                "delta_fid": get_appr_delta_metric(model_name, fid_score, num_prompts, score_type="fid"),
-                "accuracy_check": min(
-                    accuracy_check_fid(model_name, fid_score, num_prompts, mode="approx"),
-                    accuracy_check_clip(model_name, average_clip_score, num_prompts, mode="approx"),
-                ),
-                "accuracy_check_delta": min(
-                    accuracy_check_fid(model_name, fid_score, num_prompts, mode="delta"),
-                    accuracy_check_clip(model_name, average_clip_score, num_prompts, mode="delta"),
-                ),
-                "accuracy_check_valid": min(
-                    accuracy_check_fid(model_name, fid_score, num_prompts, mode="valid"),
-                    accuracy_check_clip(model_name, average_clip_score, num_prompts, mode="valid"),
-                ),
-            }
-        ],
+    metadata = {
+        "model_name": "sdxl-base-refiner-tp" if use_cfg_parallel else "sdxl-base-refiner",
+        "device": get_device_name(),
+        "device_vae": vae_on_device,
+        "capture_trace": capture_trace,
+        "encoders_on_device": encoders_on_device,
+        "use_cfg_parallel": use_cfg_parallel,
+        "denoising_split": denoising_split,
+        "refiner_strength": refiner_strength,
+        "refiner_aesthetic_score": refiner_aesthetic_score,
+        "refiner_negative_aesthetic_score": refiner_negative_aesthetic_score,
+        "num_inference_steps": num_inference_steps,
+        "start_from": start_from,
+        "num_prompts": num_prompts,
     }
 
-    os.makedirs(OUT_ROOT, exist_ok=True)
-    trace_flag = "with_trace" if capture_trace else "no_trace"
-    vae_flag = "device_vae" if vae_on_device else "host_vae"
-    encoders_flag = "device_encoders" if encoders_on_device else "host_encoders"
-    refiner_flag = f"refiner_{denoising_split}"
-    new_file_name = f"sdxl_base_and_refiner_test_results_{trace_flag}_{vae_flag}_{encoders_flag}_{use_cfg_parallel}_{num_inference_steps}_{refiner_flag}.json"
-    with open(f"{OUT_ROOT}/{new_file_name}", "w") as f:
-        json.dump(data, f, indent=4)
+    report_json = create_report_json(metadata, accuracy_metrics)
 
-    logger.info(f"Test results saved to {OUT_ROOT}/{new_file_name}")
-
-    with open(f"{OUT_ROOT}/{RESULTS_FILE_NAME}", "w") as f:  # this is for CI compatibility
-        json.dump(data, f, indent=4)
-    print(json.dumps(data, indent=4))
-
-    logger.info(f"Test results saved to {OUT_ROOT}/{RESULTS_FILE_NAME}")
-
-
-def sdxl_get_prompts(
-    captions_path,
-    start_from,
-    num_prompts,
-):
-    assert (
-        0 <= start_from < 5000 and start_from + num_prompts <= 5000
-    ), "start_from must be between 0 and 4999, and start_from + num_prompts must not exceed 5000."
-
-    prompts = []
-
-    if not os.path.isfile(captions_path):
-        logger.info(f"File {captions_path} not found. Downloading...")
-        os.makedirs(os.path.dirname(captions_path), exist_ok=True)
-        urllib.request.urlretrieve(COCO_CAPTIONS_DOWNLOAD_PATH, captions_path)
-        logger.info("Download complete.")
-
-    with open(captions_path, "r") as tsv_file:
-        reader = csv.reader(tsv_file, delimiter="\t")
-        next(reader)
-        for index, row in enumerate(reader):
-            if index < start_from:
-                continue
-            if index >= start_from + num_prompts:
-                break
-            prompts.append(row[2])
-
-    return prompts
+    save_report_json(report_json, metadata)
+    print(json.dumps(report_json, indent=4))
