+++ models/experimental/stable_diffusion_xl_base/tests/test_sdxl_inpaint_accuracy.py
@@ -4,29 +4,22 @@
 
 import pytest
 from models.experimental.stable_diffusion_xl_base.demo.demo_inpainting import test_demo
-from models.experimental.stable_diffusion_xl_base.utils.clip_encoder import CLIPEncoder
-import os
 from datasets import load_dataset
 from loguru import logger
-import statistics
-from models.experimental.stable_diffusion_xl_base.utils.fid_score import calculate_fid_score
 from models.experimental.stable_diffusion_xl_base.tests.test_common import (
     SDXL_L1_SMALL_SIZE,
     SDXL_TRACE_REGION_SIZE,
     SDXL_FABRIC_CONFIG,
 )
 import json
-from models.common.utility_functions import profiler
 from models.experimental.stable_diffusion_xl_base.conftest import get_device_name
-from models.experimental.stable_diffusion_xl_base.utils.clip_fid_ranges import (
-    accuracy_check_clip,
-    accuracy_check_fid,
-    get_appr_delta_metric,
-    get_model_targets,
+from models.experimental.stable_diffusion_xl_base.utils.accuracy_utils import (
+    calculate_accuracy_metrics,
+    create_report_json,
+    save_report_json,
 )
 
 test_demo.__test__ = False
-OUT_ROOT, RESULTS_FILE_NAME = "test_reports", "sdxl_test_results.json"
 MAX_N_SAMPLES, MIN_N_SAMPLES = 1260, 2
 
 
@@ -138,112 +131,27 @@ def test_accuracy_sdxl_inpaint(
         input_masks=input_masks,
     )
 
-    clip = CLIPEncoder()
+    accuracy_metrics = calculate_accuracy_metrics(images, prompts, coco_statistics_path)
 
-    clip_scores = []
-
-    for idx, image in enumerate(images):
-        clip_scores.append(100 * clip.get_clip_score(prompts[idx], image).item())
-
-    average_clip_score = sum(clip_scores) / len(clip_scores)
-
-    deviation_clip_score = "N/A"
-    fid_score = "N/A"
-
-    if num_prompts >= 2:
-        deviation_clip_score = statistics.stdev(clip_scores)
-        fid_score = calculate_fid_score(images, coco_statistics_path)
-    else:
-        logger.info("FID score is not calculated for less than 2 prompts.")
-
-    print(f"FID score: {fid_score}")
-    print(f"Average CLIP Score: {average_clip_score}")
-    print(f"Standard Deviation of CLIP Scores: {deviation_clip_score}")
-
-    avg_gen_end_to_end = profiler.get("end_to_end_generation")
-
-    model_name = "sdxl-inpaint-tp" if use_cfg_parallel else "sdxl-inpaint"
-    targets = get_model_targets(model_name)
-
-    data = {
-        "model": model_name,
-        "metadata": {
-            "model_name": model_name,
-            "device": get_device_name(),
-            "device_vae": vae_on_device,
-            "capture_trace": capture_trace,
-            "encoders_on_device": encoders_on_device,
-            "num_inference_steps": num_inference_steps,
-            "start_from": start_from,
-            "num_prompts": num_prompts,
-            "negative_prompt": negative_prompt,
-            "guidance_scale": guidance_scale,
-            "strength": strength,
-            "use_cfg_parallel": use_cfg_parallel,
-        },
-        "benchmarks_summary": [
-            {
-                "model": model_name,
-                "device": get_device_name(),
-                "avg_gen_time": avg_gen_end_to_end,
-                "target_checks": {
-                    "functional": {
-                        "avg_gen_time": targets["perf"]["functional"],
-                        "avg_gen_time_check": 2 if targets["perf"]["functional"] >= avg_gen_end_to_end else 3,
-                    },
-                    "complete": {
-                        "avg_gen_time": targets["perf"]["complete"],
-                        "avg_gen_time_check": 2 if targets["perf"]["complete"] >= avg_gen_end_to_end else 3,
-                    },
-                    "target": {
-                        "avg_gen_time": targets["perf"]["target"],
-                        "avg_gen_time_check": 2 if targets["perf"]["target"] >= avg_gen_end_to_end else 3,
-                    },
-                },
-                "average_denoising_time": profiler.get("denoising_loop"),
-                "average_vae_time": profiler.get("vae_decode"),
-                "min_gen_time": min(profiler.times["end_to_end_generation"]),
-                "max_gen_time": max(profiler.times["end_to_end_generation"]),
-                "average_encoding_time": profiler.get("encode_prompts"),
-            }
-        ],
-        "evals": [
-            {
-                "model": model_name,
-                "device": get_device_name(),
-                "average_clip": average_clip_score,
-                "deviation_clip": deviation_clip_score,
-                "delta_clip": get_appr_delta_metric(model_name, average_clip_score, num_prompts, score_type="clip"),
-                "fid_score": fid_score,
-                "delta_fid": get_appr_delta_metric(model_name, fid_score, num_prompts, score_type="fid"),
-                "accuracy_check": min(
-                    accuracy_check_fid(model_name, fid_score, num_prompts, mode="approx"),
-                    accuracy_check_clip(model_name, average_clip_score, num_prompts, mode="approx"),
-                ),
-            }
-        ],
+    metadata = {
+        "model_name": "sdxl-inpaint-tp" if use_cfg_parallel else "sdxl-inpaint",
+        "device": get_device_name(),
+        "device_vae": vae_on_device,
+        "capture_trace": capture_trace,
+        "encoders_on_device": encoders_on_device,
+        "use_cfg_parallel": use_cfg_parallel,
+        "negative_prompt": negative_prompt,
+        "guidance_scale": guidance_scale,
+        "strength": strength,
+        "num_inference_steps": num_inference_steps,
+        "start_from": start_from,
+        "num_prompts": num_prompts,
     }
 
-    os.makedirs(OUT_ROOT, exist_ok=True)
-    trace_flag = "with_trace" if capture_trace else "no_trace"
-    vae_flag = "device_vae" if vae_on_device else "host_vae"
-    encoders_flag = "device_encoders" if encoders_on_device else "host_encoders"
-    use_cfg_parallel_flag = "cfg_parallel" if use_cfg_parallel else "no_cfg_parallel"
-    new_file_name = (
-        f"sdxl_test_results_{trace_flag}_{vae_flag}_{encoders_flag}_{use_cfg_parallel_flag}_{num_prompts}.json"
-    )
-    with open(f"{OUT_ROOT}/{new_file_name}", "w") as f:
-        json.dump(data, f, indent=4)
+    report_json = create_report_json(metadata, accuracy_metrics)
 
-    logger.info(f"Test results saved to {OUT_ROOT}/{new_file_name}")
-
-    with open(
-        f"{OUT_ROOT}/{RESULTS_FILE_NAME}", "w"
-    ) as f:  # this is for CI and test_sdxl_accuracy_with_reset.py compatibility
-        json.dump(data, f, indent=4)
-
-    logger.info(f"Test results saved to {OUT_ROOT}/{RESULTS_FILE_NAME}")
-    print(json.dumps(data, indent=4))
+    save_report_json(report_json, metadata)
+    print(json.dumps(report_json, indent=4))
 
 
 def get_dataset_for_inpainting_accuracy(n_prompts: int):
